Reorganisation of the files (selecting feMove file for midptf )
Removing the new collision script since it is in the other branch
